+++ packages/calcite-components/src/components/date-picker-month/date-picker-month.tsx
@@ -76,6 +76,9 @@ export class DatePickerMonth extends LitElement {
   /** The currently active Date. */
   @property() activeDate: Date = new Date();
 
+  /** Specifies the number of calendars displayed when `range` is `true`. */
+  @property({ type: Number, reflect: true }) calendars: 1 | 2 = 2;
+
   /**
    * The DateTimeFormat used to provide screen reader labels.
    *
@@ -618,7 +621,9 @@ export class DatePickerMonth extends LitElement {
     return (
       <div class={{ [CSS.calendarContainer]: true }} role="grid">
         {this.renderCalendar(adjustedWeekDays, days)}
-        {this.range && this.renderCalendar(adjustedWeekDays, nextMonthDays, true)}
+        {this.range &&
+          this.calendars === 2 &&
+          this.renderCalendar(adjustedWeekDays, nextMonthDays, true)}
       </div>
     );
   }
@@ -696,7 +701,7 @@ export class DatePickerMonth extends LitElement {
           min={this.min}
           monthStyle={this.monthStyle}
           oncalciteInternalDatePickerMonthHeaderSelectChange={this.monthHeaderSelectChange}
-          position={isEndCalendar ? "end" : this.range ? "start" : null}
+          position={isEndCalendar ? "end" : this.range && this.calendars === 2 ? "start" : null}
           scale={this.scale}
           selectedDate={this.selectedDate}
         />