+++ packages/calcite-components/src/components/input-date-picker/input-date-picker.tsx
@@ -190,6 +190,9 @@ export class InputDatePicker
 
   //#region Public Properties
 
+  /** Specifies the number of calendars displayed when `range` is `true`. */
+  @property({ type: Number, reflect: true }) calendars: 1 | 2 = 2;
+
   /** When `true`, interaction is prevented and the component is displayed with lower opacity. */
   @property({ reflect: true }) disabled = false;
 
@@ -906,6 +909,11 @@ export class InputDatePicker
     if (this.shouldFocusRangeStart() || this.rangeStartValueChangedByUser) {
       return;
     }
+
+    if (this.proximitySelectionDisabled && this.valueAsDate[1] === null) {
+      return;
+    }
+
     this.open = false;
     this.focusInput();
   }
@@ -1142,6 +1150,7 @@ export class InputDatePicker
                 <calcite-date-picker
                   activeDate={this.datePickerActiveDate}
                   activeRange={this.focusedInput}
+                  calendars={this.calendars}
                   headingLevel={this.headingLevel}
                   layout={this.layout}
                   max={this.max}