+++ packages/calcite-components/src/components/input/input.stories.ts
@@ -260,3 +260,15 @@ export const fontSizeSetAtRoot = (): string =>
     <calcite-input placeholder="Placeholder" prefix-text="Prefix" suffix-text="Suffix" type="text" icon="search">
       <calcite-button slot="action"> Search </calcite-button>
     </calcite-input>`;
+
+export const overlayDoesNotObscureIcon = (): string =>
+  html` <style>
+      .overlay {
+        position: absolute;
+        inset: 0;
+        background-color: white;
+        opacity: 0.75;
+      }
+    </style>
+    <calcite-input icon="check-square-f"></calcite-input>
+    <div class="overlay"></div>`;