+++ packages/calcite-components/src/components/list-item/list-item.tsx
@@ -12,7 +12,7 @@ import { SelectionMode, InteractionMode, Scale, FlipContext } from "../interface
 import { SelectionAppearance } from "../list/resources";
 import { IconNameOrString } from "../icon/interfaces";
 import { SortableComponentItem } from "../../utils/sortableComponent";
-import { MoveTo } from "../sort-handle/interfaces";
+import { SortMenuItem } from "../sort-handle/interfaces";
 import { useT9n } from "../../controllers/useT9n";
 import type { SortHandle } from "../sort-handle/sort-handle";
 import type { List } from "../list/list";
@@ -177,11 +177,18 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
   @property({ reflect: true }) displayMode: ListDisplayMode = "flat";
 
   /**
-   * Sets the item to display a border.
+   * Defines the "Add to" items.
    *
    * @private
    */
-  @property() moveToItems: MoveTo[] = [];
+  @property() addToItems: SortMenuItem[] = [];
+
+  /**
+   * Defines the "Move to" items.
+   *
+   * @private
+   */
+  @property() moveToItems: SortMenuItem[] = [];
 
   /**
    * When `true`, the item is open to show child components.
@@ -330,12 +337,6 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
    * @private
    */
   calciteInternalListItemToggle = createEvent({ cancelable: false });
-
-  /**
-   *
-   * @private
-   */
-  calciteInternalListItemUpdateMoveToItems = createEvent({ cancelable: false });
 
   /** Fires when the close button is clicked. */
   calciteListItemClose = createEvent({ cancelable: false });
@@ -487,7 +488,6 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
   private handleSortHandleBeforeOpen(event: CustomEvent<void>): void {
     event.stopPropagation();
     this.calciteListItemSortHandleBeforeOpen.emit();
-    this.calciteInternalListItemUpdateMoveToItems.emit();
   }
 
   private handleSortHandleBeforeClose(event: CustomEvent<void>): void {
@@ -770,8 +770,16 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
   }
 
   private renderDragHandle(): JsxNode {
-    const { label, dragHandle, dragDisabled, setPosition, setSize, moveToItems, sortDisabled } =
-      this;
+    const {
+      label,
+      dragHandle,
+      dragDisabled,
+      setPosition,
+      setSize,
+      moveToItems,
+      sortDisabled,
+      addToItems,
+    } = this;
 
     return dragHandle ? (
       <div
@@ -782,6 +790,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
         role="gridcell"
       >
         <calcite-sort-handle
+          addToItems={addToItems}
           disabled={dragDisabled}
           label={label}
           moveToItems={moveToItems}